Sun of numbers in string
Problem Statement 1: 

Given a string, return the sum of the digits 0-9 that appear in the string, ignoring all other characters. A number is a series of 1 or more digit chars in a row. Return 0 if there are no digits in the string.

 Example 1: 
Input: aa1bc2d3
Output: 6
Explanation: The string contains three digits such as 1,2 and 3, sum of the digits 6

 Example 2: 
Input: Chocolate
Output: 0
Explanation: The string does not contains any digits, sum of the digits 0

 Example 3: 
Input: aa11b33
Output: 44
Explanation: The string contains two digits such as 11 and 33, sum of the digits 44